requirements.txt
streamlit
google-genai
gspread
oauth2client
pandas